requirements.txt
streamlit>=1.30.0
bcrypt>=4.0.0
google-genai>=1.0.0
together>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
streamlit-mermaid>=0.2.0
